Expanded scripts to work with Ubuntu 14.04 and to fully install unixODBC
A key change was to add the new function install_dm. It seemed silly
for a script to require a manual step. As a precaution, moved the
installation warning to the beginning and created a prompt to get
confirmation before proceeding.

Included the CPPFLAGS as recommended by Microsoft in their
documentation. And exposed the —accept-warning flag in the help display.